// frontend/src/api/models/Semester.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Semester = {
    First: 'FIRST',
    Second: 'SECOND',
    Fourth: 'FOURTH',
    Sixth: 'SIXTH',
    Seventh: 'SEVENTH'
} as const;
export type Semester = typeof Semester[keyof typeof Semester];


export function instanceOfSemester(value: any): boolean {
    for (const key in Semester) {
        if (Object.prototype.hasOwnProperty.call(Semester, key)) {
            if (Semester[key as keyof typeof Semester] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SemesterFromJSON(json: any): Semester {
    return SemesterFromJSONTyped(json, false);
}

export function SemesterFromJSONTyped(json: any, ignoreDiscriminator: boolean): Semester {
    return json as Semester;
}

export function SemesterToJSON(value?: Semester | null): any {
    return value as any;
}

export function SemesterToJSONTyped(value: any, ignoreDiscriminator: boolean): Semester {
    return value as Semester;
}