// frontend/src/api/models/StudyGroupExpelledTotalResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StudyGroupExpelledTotalResponse
 */
export interface StudyGroupExpelledTotalResponse {
    /**
     * 
     * @type {number}
     * @memberof StudyGroupExpelledTotalResponse
     */
    totalExpelledStudents: number;
}

/**
 * Check if a given object implements the StudyGroupExpelledTotalResponse interface.
 */
export function instanceOfStudyGroupExpelledTotalResponse(value: object): value is StudyGroupExpelledTotalResponse {
    if (!('totalExpelledStudents' in value) || value['totalExpelledStudents'] === undefined) return false;
    return true;
}

export function StudyGroupExpelledTotalResponseFromJSON(json: any): StudyGroupExpelledTotalResponse {
    return StudyGroupExpelledTotalResponseFromJSONTyped(json, false);
}

export function StudyGroupExpelledTotalResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudyGroupExpelledTotalResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'totalExpelledStudents': json['totalExpelledStudents'],
    };
}

export function StudyGroupExpelledTotalResponseToJSON(json: any): StudyGroupExpelledTotalResponse {
    return StudyGroupExpelledTotalResponseToJSONTyped(json, false);
}

export function StudyGroupExpelledTotalResponseToJSONTyped(value?: StudyGroupExpelledTotalResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'totalExpelledStudents': value['totalExpelledStudents'],
    };
}