// frontend/src/api/models/StudyGroupShouldBeExpelledGroupResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IS Project API
 * Бэкенд для управления учебными группами и связанными сущностями. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StudyGroupShouldBeExpelledGroupResponse
 */
export interface StudyGroupShouldBeExpelledGroupResponse {
    /**
     * 
     * @type {number}
     * @memberof StudyGroupShouldBeExpelledGroupResponse
     */
    shouldBeExpelled: number;
    /**
     * 
     * @type {number}
     * @memberof StudyGroupShouldBeExpelledGroupResponse
     */
    count: number;
}

/**
 * Check if a given object implements the StudyGroupShouldBeExpelledGroupResponse interface.
 */
export function instanceOfStudyGroupShouldBeExpelledGroupResponse(value: object): value is StudyGroupShouldBeExpelledGroupResponse {
    if (!('shouldBeExpelled' in value) || value['shouldBeExpelled'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    return true;
}

export function StudyGroupShouldBeExpelledGroupResponseFromJSON(json: any): StudyGroupShouldBeExpelledGroupResponse {
    return StudyGroupShouldBeExpelledGroupResponseFromJSONTyped(json, false);
}

export function StudyGroupShouldBeExpelledGroupResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudyGroupShouldBeExpelledGroupResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'shouldBeExpelled': json['shouldBeExpelled'],
        'count': json['count'],
    };
}

export function StudyGroupShouldBeExpelledGroupResponseToJSON(json: any): StudyGroupShouldBeExpelledGroupResponse {
    return StudyGroupShouldBeExpelledGroupResponseToJSONTyped(json, false);
}

export function StudyGroupShouldBeExpelledGroupResponseToJSONTyped(value?: StudyGroupShouldBeExpelledGroupResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'shouldBeExpelled': value['shouldBeExpelled'],
        'count': value['count'],
    };
}